+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/FeatureGates.java
@@ -16,11 +16,9 @@ import static java.util.Arrays.asList;
 public class FeatureGates {
     /* test */ static final FeatureGates NONE = new FeatureGates("");
 
-    private static final String USE_KRAFT = "UseKRaft";
     private static final String CONTINUE_ON_MANUAL_RU_FAILURE = "ContinueReconciliationOnManualRollingUpdateFailure";
 
     // When adding new feature gates, do not forget to add them to allFeatureGates() and toString() methods
-    private final FeatureGate useKRaft = new FeatureGate(USE_KRAFT, true);
     private final FeatureGate continueOnManualRUFailure =
         new FeatureGate(CONTINUE_ON_MANUAL_RU_FAILURE, false);
 
@@ -44,9 +42,6 @@ public class FeatureGates {
                 featureGate = featureGate.substring(1);
 
                 switch (featureGate) {
-                    case USE_KRAFT:
-                        setValueOnlyOnce(useKRaft, value);
-                        break;
                     case CONTINUE_ON_MANUAL_RU_FAILURE:
                         setValueOnlyOnce(continueOnManualRUFailure, value);
                         break;
@@ -84,13 +79,6 @@ public class FeatureGates {
     }
 
     /**
-     * @return  Returns true when the UseKRaft feature gate is enabled
-     */
-    public boolean useKRaftEnabled() {
-        return useKRaft.isEnabled();
-    }
-
-    /**
      * @return  Returns true when the ContinueReconciliationOnManualRollingUpdateFailure feature gate is enabled
      */
     public boolean continueOnManualRUFailureEnabled() {
@@ -103,16 +91,12 @@ public class FeatureGates {
      * @return  List of all Feature Gates
      */
     /*test*/ List<FeatureGate> allFeatureGates()  {
-        return List.of(
-                useKRaft,
-            continueOnManualRUFailure
-        );
+        return List.of(continueOnManualRUFailure);
     }
 
     @Override
     public String toString() {
         return "FeatureGates(" +
-                "UseKRaft=" + useKRaft.isEnabled() +
                 "ContinueReconciliationOnManualRollingUpdateFailure=" + continueOnManualRUFailure.isEnabled() +
                 ")";
     }